2025/requirements.txt
openai>=1.0.0
anthropic>=0.3.0
schedule>=1.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
pytz>=2024.1
